+++ oci/aws-cli/v2/install.aws-cli-v2.sh
@@ -17,17 +17,18 @@ set +x
 export HORODATAGE=`date +%m-%d-%Y_%Hh-%Mmin-%Ssec`
 export OPS_HOME=$(pwd)
 
-# ------
-# Ensuring that the paths you install to contain no
-# volume or folder names that contain any spaces otherwise
-# the installation would fail.
-mkdir -p ~/.a-k8s-demo/aws_cli
-export PROVISIONING_HOME=$(mktemp  --tmpdir=/home/${USER}/.a-k8s-demo/aws_cli -d -t provisioning.${HORODATAGE}.XXX)
+
 # unused yet, AWS releases only major versions
 export AWS_CLI_VERSION=${AWS_CLI_VERSION:-'0.0.0'}
 export AWS_CLI_MAJOR_VERSION=${AWS_CLI_MAJOR_VERSION:-'2'}
 export AWS_CLI_PACKAGE_DWNLD_URI=${AWS_CLI_PACKAGE_DWNLD_URI:-'https://awscli.amazonaws.com/awscli-exe-linux-x86_64.zip'}
 export AWS_CLI_PACKAGE_FILENAME=awscliv2.zip
+# ------
+# Ensuring that the paths you install to contain no
+# volume or folder names that contain any spaces otherwise
+# the installation would fail.
+mkdir -p /home/$(whoami)/.a-k8s-demo/aws_cli/${AWS_CLI_VERSION}/
+export PROVISIONING_HOME=$(mktemp -d -p /home/$(whoami)/.a-k8s-demo/aws_cli/${AWS_CLI_VERSION}/ -t provision.XXXXXXX)
 
 echo ""
 echo "Downloading [$AWS_CLI_PACKAGE_FILENAME] ..."
@@ -45,10 +46,4 @@ ls -allh $PROVISIONING_HOME/aws
 
 $PROVISIONING_HOME/aws/install
 
-if [ $? == 0 ]; then
-  aws --version
-  aws iam list-users
-else
-  echo "Installing AWS CLI v2 failed, mzybe because you are not executing this as root"
-  exit 1
-fi;
+aws --version