+++ oci/kubeone_operator/terraform.minikube/main.tf
@@ -17,7 +17,8 @@ resource "aws_instance" "creshVM" {
   ami             = var.aws_instance_desired_ami #id of desired AMI
   instance_type   = var.aws_instance_type
   # keypair is created using AWS CLI : aws ec2 create-key-pair --key-name creshKeyPair --query 'KeyMaterial' --output text > ./aws.creshkey.pem
-  key_name = "creshKeyPair"
+  # key_name = "creshKeyPair"
+  key_name = "jblCreshPaireClef"
   # key_name = "${module.aws_key_pair.deployercreds.key_name}"
   security_groups = ["${aws_security_group.allow_all.name}"]
   # iam_instance_profile = "${aws_iam_instance_profile.test_profile.name}"
@@ -44,13 +45,31 @@ resource "aws_eip_association" "eip_assoc" {
 resource "aws_security_group" "allow_all" {
   name = "allow_ssh"
   ingress {
-    from_port
-    to_port
+    from_port   = "0"
+    to_port     = "0"
     protocol    = "-1"
     cidr_blocks = ["0.0.0.0/0"]
   }
+
+  # ---
+  # outbound traffic to at least, be able to
+  # use package manager inside the
+  # linux amazon AMI distro
+  # ---
+  egress {
+  from_port   = 0
+  to_port     = 0
+  protocol    = "-1"
+  cidr_blocks = ["0.0.0.0/0"]
+  }
 }
+
+
+# ---
+# Test it : [terraform import aws_key_pair.deployercreds creshkey]
+# ---
 # resource "aws_key_pair" "deployercreds" {
 #   key_name   = "creshkey"
 #   public_key = "var.my_ssh_pubkey"
 # }
+#